Add --initial-best-value option
The default value is 0.5. This is the best value for the initial parameter values. In the first trial or trial 0, the test engine will take the suggested parameter values from optimizer and the base engine will take the initial parameter values. After the game match between the test engine and the base engine, if the score of test engine could not exceed the initial_best_value then the best  value will be the initial_best_value and the best parameter values will be the initial parameter values.

This option is important when you want to run a smaller number of games per trial but would like to have a high confidence of the tuning. You can use for example --initial-best-value 0.75, with --games-per-trial 8. That means that the suggested parameter values from optimizer needs to score 6/8 in order for it  to become the best parameter and replace the default parameter values. The reliability of optimization result, depends on the games-per-trial value (high is better) and the initial-best-value (high is better and above 0.5). The optimization result is verified by running many game matches between the default and the best parameter values output.