Expose consider_pruned_trials param in skopt sampler
Optuna framework by default sets consider_pruned_trials = [True or False] of skopt or scikit-optimize sampler to false because the sampling may take some time as trials increases. But in our case where every trial may take time and when pruning is activated the pruned trials has to be considered in the sampling.

When using skopt, the recommended param command is:
--sampler name=skopt consider_pruned_trials=true ...