Modify skopt initial_point_generator to sobol sequence
In the scikit optimize library the default is 'random'. The issue with random is that in the first 10 trials the suggested param values were not randomized (the param values suggested are the same for each trial) when random_state is not None. With sobol sequence the param values in the first 10 trials are different which is what we like it to be.

Sobol sequence ref: https://en.wikipedia.org/wiki/Sobol_sequence

scikit optimize has many initial_point_generator methods, see ref.: https://scikit-optimize.github.io/stable/modules/generated/skopt.optimizer.Optimizer.html

At the moment initial point generator for skopt sampler is not settable by the user.