Remove --fix-base-param flag, this is now the default behavior
* Always use the default or init param values specified by the user for the base engine.
* Samplers or optimizers may repeat suggesting the same param values (specially when the number of param is low and the range of param values is also low and when such params produces a good objective value signalling convergence to optimal values) from trial to trial. In this case the dynamic base param scheme is ineffective. So for now the base param used will be always the init or default param as specified by the user.

In my todo list, I will have to handle the situation when sampler attempts to suggest params that were are already tried in the past. I can ignore it and asks for a new param again for x times or so. Or accept the repeated param suggestion and give the old value (without doing engine vs engine matches) that was already in the trial history.