Add deterministic function info in the log
Example:

starting trial: 66 ...
deterministic function: False
suggested param for test engine: {'PawnValueEn': 110, 'PawnValueOp': 90}
param for base engine          : {'PawnValueEn': 125, 'PawnValueOp': 95}

When --noisy-result flag is enabled, deterministic function is false.

By default deterministic function is true meaning when the optimizer repeats a param the tuner will just report back the old objective value from the old param that was tried before.

When the --noisy-result flag is enabled and when the optimizer repeats the same param it will play the match and the objective value might not be the same from the old param that was tried before. In this case the tuner will calculate the mean objective value for all same param and it also records how many times a param has been repeated in the log_tuner.txt.

An example of using --noisy-result flag is when your games per trial is low and you are using a time control. As an experiment, try to run engine vs engine match for 50 games at TC=1s+50ms. Record the match result say 0.55 for first engine. Now repeat the same match and record the result. If result is not 0.55 or not close to it, say 0.45, you better use --noisy-result flag.

If you play at fixed depth say depth 5 for 50 games and result of both match is close like 0.55 and 0.57, do not use --noisy-result flag. In this case the tuner will just return 0.55 without playing games.